Split e2e tests
