+++ azkustodata/test/.gitignore
@@ -1,0 +1,3 @@
+config.json
+*.csv
+!testdata/*.csv
